refactor DelegationModal and related components; update import paths and add new styles
