+++ src/features/amUI/common/DelegationModal/AccessPackages/PackageSearch.tsx
@@ -2,6 +2,10 @@ import { Trans, useTranslation } from 'react-i18next';
 import { Heading, Search } from '@digdir/designsystemet-react';
 import { useState } from 'react';
 
+import { useDelegationModalContext } from '../DelegationModalContext';
+
+import classes from './PackageSearch.module.css';
+
 import { debounce } from '@/resources/utils';
 import type { AccessPackage } from '@/rtk/features/accessPackageApi';
 import type { Party } from '@/rtk/features/lookupApi';
