+++ src/features/amUI/common/DelegationModal/SingleRights/ResourceInfo.module.css
