+++ src/features/amUI/common/DelegationModal/SingleRights/ResourceSearch.module.css
