+++ src/features/amUI/reporteeRightsPage/RoleInfoModal.tsx
@@ -1,7 +1,5 @@
 import { useEffect } from 'react';
 
-import { EditModal } from '../../userRightsPage/DelegationModal/EditModal';
-
 import type { Party } from '@/rtk/features/lookupApi';
 import type { Role } from '@/rtk/features/roleApi';
